Began seperation of game and solver
